// src/redux/filter/types.ts
export enum SortPropertyEnum {
	RAITING_DESC = 'rating',
	RAITING_ASC = '-rating',
	TITLE_DESC = 'title',
	TITLE_ASC = '-title',
	PRICE_DESC = 'price',
	PRICE_ASC = '-price',
}

export type Sort = {
	name: string
	sortProperty: SortPropertyEnum
}

export interface FilterSliceState {
	searchValue: string
	categoryId: number
	currentPage: number
	sort: Sort
}